requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
# ML Libraries (for Kaggle dataset training)
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
joblib==1.3.1
nltk==3.8.1
# Optional: For advanced transformer models
# transformers==4.33.0
# torch==2.0.1
# Optional: For Kaggle dataset download
# kaggle==1.5.13
